fixup msp4th on PC testing

+++ ns430.h
@@ -3,8 +3,8 @@
 
 typedef unsigned char uint8_t;
 typedef signed char int8_t;
-typedef unsigned int uint16_t;
-typedef signed int int16_t;
+typedef unsigned short uint16_t;
+typedef signed short int16_t;
 
 typedef __attribute__((aligned (2))) uint8_t str_t;
